fix camerashake (#377)
* 修复npc-api大量bug
实现滚动字幕api

* 尝试解决内存泄漏问题

* 使用google guava解决客户端不响应导致的内存泄漏问题

* 修正

* 尝试修复出生点问题

* 修复 Player.shakeCamera() 方法中的一个错误参数

+++ src/main/java/cn/nukkit/Player.java
@@ -6669,7 +6669,7 @@ public class Player extends EntityHuman implements CommandSender, InventoryHolde
 
     @PowerNukkitXOnly
     @Since("1.6.0.0-PNX")
-    public void shakeCamera(float intensity, int duration, CameraShakePacket.CameraShakeType shakeType, CameraShakePacket.CameraShakeAction shakeAction) {
+    public void shakeCamera(float intensity, float duration, CameraShakePacket.CameraShakeType shakeType, CameraShakePacket.CameraShakeAction shakeAction) {
         CameraShakePacket packet = new CameraShakePacket();
         packet.intensity = intensity;
         packet.duration = duration;
